+++ magik-language-server/src/test/java/nl/ramsolutions/sw/magik/languageserver/inlayhint/InlayHintProviderTest.java
@@ -65,7 +65,7 @@ class InlayHintProviderTest {
                     ParameterDefinition.Modifier.NONE,
                     TypeString.UNDEFINED)),
             null,
-            Collections.emptySet(),
+            null,
             ExpressionResultString.UNDEFINED,
             ExpressionResultString.EMPTY));